refactor: (review) merge board list view and row view
